refactor: organise get_embeddings JS logic into folders
