+++ nodes/web/helpers/get_embeddings/main.js
@@ -1,0 +1,3 @@
+import { updateInputsOutputs } from "./utils.js";
+
+export { updateInputsOutputs };
